Moved DetailsModal and some logic up (not technically neccessary but it felt wrong to  have DetailsModal  placed under Sidebar
  Use  useDetailsModalState in DetailsModal
